requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
PyQt5>=5.15.0
opencv-python>=4.5.0
astropy>=5.0.0
scikit-image>=0.19.0
scipy>=1.7.0
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
pytest>=6.2.0
hypothesis>=6.0.0
pytest-cov>=2.12.0
